refactor: move SendState and ReceiveState to commons
Move wallet SendState and ReceiveState sealed classes from
amethyst.ui.screen.loggedIn.wallet to commons.ui.wallet for KMP.
